feat: Add GCD of strings algorithm
The code changes introduce a new algorithm for finding the greatest common divisor (GCD) of two strings. The algorithm calculates the GCD string by determining the greatest common divisor of the lengths of the two strings and checking if the string is a common divisor of both strings.

This addition enhances the functionality of the codebase by providing a convenient way to find the GCD of strings.

References:
- [LeetCode](https://leetcode.com/problems/greatest-common-divisor-of-strings/)
- [GeeksForGeeks](https://www.geeksforgeeks.org/gcd-of-two-strings/)