feat: Add implementation for the productExceptSelf algorithm
This commit adds the implementation for the productExceptSelf algorithm, which takes an array of integers and returns an array where each element is the product of all the elements in the input array except itself. The algorithm solves the problem without using division and has a time complexity of O(n). The implementation includes two versions of the algorithm: productExceptSelf and productExceptSelfV2. The code also includes unit tests for both versions of the algorithm.

References:
- [LeetCode Problem](https://leetcode.com/problems/product-of-array-except-self/)
- [ProgramCreek](https://www.programcreek.com/2014/07/leetcode-product-of-array-except-self-java/)